# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  apperrors.APIErrorResponse:
    description: Error response object
    properties:
      code:
        description: Application-specific error code
        example: 400
        type: integer
      message:
        description: Human-readable message for clients
        example: invalid request body
        type: string
    type: object
  model.Address:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      state:
        type: string
      zip:
        type: string
    type: object
  model.Customer:
    properties:
      address:
        type: string
      city:
        type: string
      company:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      firstName:
        maxLength: 100
        type: string
      id:
        type: integer
      lastName:
        maxLength: 100
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      org:
        $ref: '#/definitions/model.Org'
      orgId:
        type: integer
      phoneNumber:
        type: string
      state:
        type: string
      updated_at:
        type: string
    required:
    - firstName
    - lastName
    type: object
  model.DeliveryInfo:
    properties:
      address:
        type: string
      at:
        type: string
      city:
        type: string
      country:
        type: string
      date:
        type: string
      state:
        type: string
      status:
        $ref: '#/definitions/model.DeliveryStatus'
      transportFare:
        type: number
    type: object
  model.DeliveryStatus:
    enum:
    - pending
    - shipped
    - delivered
    - cancelled
    type: string
    x-enum-varnames:
    - DeliveryPending
    - DeliveryShipped
    - DeliveryDelivered
    - DeliveryCancelled
  model.DiscountInfo:
    properties:
      amount:
        type: number
      type:
        $ref: '#/definitions/model.DiscountType'
    type: object
  model.DiscountType:
    enum:
    - percentage
    - fixed
    type: string
    x-enum-varnames:
    - DiscountPercentage
    - DiscountFixed
  model.Order:
    properties:
      created_at:
        type: string
      createdBy:
        $ref: '#/definitions/model.User'
      customerId:
        type: integer
      delivery:
        $ref: '#/definitions/model.DeliveryInfo'
      discount:
        $ref: '#/definitions/model.DiscountInfo'
      distributorId:
        type: integer
      id:
        type: integer
      notes:
        type: string
      orderItems:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      org:
        $ref: '#/definitions/model.Org'
      orgId:
        type: integer
      requestedFor:
        $ref: '#/definitions/model.Customer'
      status:
        $ref: '#/definitions/model.OrderStatus'
      updated_at:
        type: string
    type: object
  model.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order:
        $ref: '#/definitions/model.Order'
      orderId:
        type: integer
      price:
        type: number
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  model.OrderStatus:
    enum:
    - pending
    - approved
    - delivered
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusApproved
    - OrderStatusDelivered
    - OrderStatusCancelled
  model.Org:
    description: Organization response model
    properties:
      address:
        $ref: '#/definitions/model.Address'
      created_at:
        type: string
      customers:
        items:
          $ref: '#/definitions/model.Customer'
        type: array
      email:
        maxLength: 50
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        maxLength: 50
        type: string
      onboardedAt:
        type: boolean
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      organizationName:
        maxLength: 50
        type: string
      organizationUrl:
        maxLength: 100
        type: string
      phone:
        maxLength: 50
        type: string
      products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/model.Customer'
        type: array
    required:
    - email
    - name
    - organizationName
    - phone
    type: object
  model.Product:
    description: Product response model
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      name:
        type: string
      org:
        $ref: '#/definitions/model.Org'
      orgId:
        type: integer
      updated_at:
        type: string
      variants:
        items:
          $ref: '#/definitions/model.Variant'
        type: array
    type: object
  model.Role:
    enum:
    - distributor
    - admin
    - viewer
    type: string
    x-enum-varnames:
    - RoleOwner
    - RoleAdmin
    - RoleViewer
  model.User:
    description: User response model
    properties:
      address:
        $ref: '#/definitions/model.Address'
      clerkId:
        type: string
      created_at:
        type: string
      email:
        maxLength: 50
        type: string
      firstName:
        maxLength: 100
        type: string
      id:
        type: integer
      lastName:
        maxLength: 100
        type: string
      org:
        allOf:
        - $ref: '#/definitions/model.Org'
        description: optional relation
      orgId:
        description: nullable foreign key
        type: integer
      phone:
        maxLength: 50
        type: string
      role:
        $ref: '#/definitions/model.Role'
      updated_at:
        type: string
    required:
    - email
    - firstName
    - lastName
    type: object
  model.Variant:
    description: Variant response model
    properties:
      color:
        type: string
      id:
        type: integer
      orgID:
        description: needed for sku uniqueness per org
        type: integer
      price:
        type: number
      productID:
        type: integer
      size:
        type: string
      sku:
        type: string
      stock:
        type: integer
    type: object
...
  product.createProductDTO:
    properties:
      category:
        maxLength: 50
        minLength: 2
        type: string
      description:
        maxLength: 1000
        minLength: 2
        type: string
      imageUrl:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      variants:
        items:
          $ref: '#/definitions/product.createProductVariantDTO'
        type: array
    required:
    - name
    - variants
    type: object
  product.createProductVariantDTO:
    properties:
      color:
        maxLength: 30
        minLength: 1
        type: string
      price:
        type: number
      size:
        maxLength: 30
        minLength: 1
        type: string
      sku:
        maxLength: 50
        minLength: 2
        type: string
      stock:
        minimum: 0
        type: integer
    required:
    - price
    - sku
    - stock
    type: object
  product.createVariantDTO:
    properties:
      color:
        maxLength: 30
        minLength: 1
        type: string
      price:
        type: number
      size:
        maxLength: 30
        minLength: 1
        type: string
      sku:
        maxLength: 50
        minLength: 2
        type: string
      stock:
        minimum: 0
        type: integer
    required:
    - price
    - sku
    - stock
    type: object
  product.updateProductDTO:
    properties:
      category:
        maxLength: 50
        minLength: 2
        type: string
      description:
        maxLength: 1000
        minLength: 2
        type: string
      imageUrl:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
  product.updateVariantDTO:
    properties:
      color:
        maxLength: 30
        minLength: 1
        type: string
      price:
        type: number
      size:
        maxLength: 30
        minLength: 1
        type: string
      stock:
        minimum: 0
        type: integer
    type: object
  response.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: success
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@openb2b.com
    name: API Support
    url: http://openb2b.com/support
  description: Open-source multi-tenant ordering & invoicing platform API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://openb2b.com/terms/
  title: OpenB2B API
  version: "1.0"
paths:
...
  /products:
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.createProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create products
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - products
    get:
      description: Get a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update product payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.updateProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - products
  /products/{id}/variants/{id}:
    delete:
      description: Delete a variant by product ID and variant ID
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete variant
      tags:
      - variants
    get:
      description: Get a variant by product ID and variant ID
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      - description: variant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Variant'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get variant
      tags:
      - variants
  /products/{productId}/variants:
    post:
      consumes:
      - application/json
      description: Create a new variant
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      - description: Variant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.createVariantDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Variant'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create variant
      tags:
      - variants
  /products/{productId}/variants/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing variant by product ID and variant ID
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update variant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.updateVariantDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Variant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update variant
      tags:
      - variants
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"